Trait archisinal_lib::traits::account_manager::AccountManager

pub trait AccountManager {
    // Required methods
    fn create_account(&mut self) -> ProjectResult<()>;
    fn create_creator_account(&mut self) -> ProjectResult<()>;
    fn get_account(&self, account_id: AccountId) -> Option<AccountId>;
    fn get_creator_account(&self, account_id: AccountId) -> Option<AccountId>;
    fn get_creator_code_hash(&self) -> Hash;
    fn get_user_code_hash(&self) -> Hash;
    fn set_creator_code_hash(&mut self, code_hash: Hash) -> ProjectResult<()>;
    fn set_user_code_hash(&mut self, code_hash: Hash) -> ProjectResult<()>;
}

Account manager trait

This trait is used to manage accounts and creators

Required Methods

fn create_account(&mut self) -> ProjectResult<()>

Create an account

This function will deploy an account contract for the caller using the user code hash

Errors

ArchisinalError::AccountAlreadyExists - If the account already exists

Returns

ProjectResult<()> - Ok if the account was created, otherwise an error

Emits

AccountCreated - If the account was created successfully

fn create_creator_account(&mut self) -> ProjectResult<()>

Create a creator account

This function will deploy a creator account contract for the caller using the creator code hash

Errors

ArchisinalError::AccountAlreadyExists - If the account already exists

Returns

ProjectResult<()> - Ok if the account was created, otherwise an error

Emits

CreatorAccountCreated - If the creator account was created successfully

fn get_account(&self, account_id: AccountId) -> Option<AccountId>

Get an account

This function will get an account contract address

Arguments

account_id - An address of the user, for which to get the account contract address

Returns

Option<AccountId> - An account contract address if it exists, otherwise None

fn get_creator_account(&self, account_id: AccountId) -> Option<AccountId>

Get a creator account

This function will get a creator account contract address

Arguments

account_id - An address of the user, for which to get the creator contract address

Returns

Option<AccountId> - A creator account contract address if it exists, otherwise None

fn get_creator_code_hash(&self) -> Hash

Get the creator code hash

This function will get the creator code hash, which is used to deploy creator accounts

Returns

Hash - The creator code hash

fn get_user_code_hash(&self) -> Hash

Get the user code hash

This function will get the user code hash, which is used to deploy user accounts

Returns

Hash - The user code hash

fn set_creator_code_hash(&mut self, code_hash: Hash) -> ProjectResult<()>

Set the creator code hash

This function will set the creator code hash, which is used to deploy creator accounts

Arguments

code_hash - The creator code hash

Returns

ProjectResult<()> - Ok if the code hash was set, otherwise an error

Emits

CreatorCodeHashSet - If the creator code hash was set successfully

fn set_user_code_hash(&mut self, code_hash: Hash) -> ProjectResult<()>

Set the user code hash

This function will set the user code hash, which is used to deploy user accounts

Arguments

code_hash - The user code hash

Returns

ProjectResult<()> - Ok if the code hash was set, otherwise an error

Emits

UserCodeHashSet - If the user code hash was set successfully